requirements.txt
openai==1.35.13
prompt_toolkit==3.0.47
python-dotenv==1.0.1
